Add InstanceDescriptor directly to the ExecutionContext

+++ exonum/src/runtime/error/core_errors.rs
@@ -60,7 +60,7 @@ pub enum CoreError {
 }
 
 impl CoreError {
-    pub(crate) fn stack_overflow(max_depth: usize) -> ExecutionError {
+    pub(crate) fn stack_overflow(max_depth: u64) -> ExecutionError {
         let description = format!(
             "Maximum depth of call stack ({}) has been reached.",
             max_depth
